app/requirements.txt
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.1
gunicorn==19.9.0
Markdown==3.2.2
